% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrad.R
\name{deg2rad}
\alias{deg2rad}
\alias{rad2deg}
\title{Function for transforming deg2rad and rad2deg}
\usage{
deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{deg}{numerical}

\item{rad}{numerical}
}
\value{
numeric value
}
\description{
Function for transforming deg2rad and rad2deg
}
\examples{
deg2rad (30)
rad2deg (pi/2)

}
