% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist3d.R
\name{dist3d}
\alias{dist3d}
\title{Function for computing distance in 3D space from point p to line described by vector u attached in point p0}
\usage{
dist3d(p = c(1, 1, 1), u = c(1, 1, 1), p0 = c(0, 0, 0))
}
\arguments{
\item{p}{numeric 3 element vector describing selected point to measure distance to}

\item{u}{numeric 3 element vector describing direction of the line of sight}

\item{p0}{numeric 3 element vector describing point of attachment of vector u}
}
\value{
Numeric distance is returned.
}
\description{
Function for computing distance in 3D space from point p to line described by vector u attached in point p0
}
\examples{
dist3d(p,u,p0)

}
