% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dronePos.R
\name{dronePos}
\alias{dronePos}
\title{Function for generating drone position in tine and 3D space}
\usage{
dronePos(
  time,
  xfun = function(time) 100 + sin(time),
  yfun = function(time) 20 + 3 * sin(time),
  zfun = function(time) 80 + 1 * sin(time),
  timeunits = "s",
  dimunits = "m"
)
}
\arguments{
\item{time}{time vector e.g. t= 1:100}

\item{xfun}{yfun, zfun - functions generating points in time}
}
\value{
return an array of x,y,z positions of drone
}
\description{
Function for generating drone position in tine and 3D space
}
\details{
an example of funciton: xfun = function(time) 5 + 2*sin(time)
}
\examples{
dronePos(t,xfun,yfun,zfun)

}
