% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{Rx}
\alias{Rx}
\alias{Ry}
\alias{Rz}
\alias{Raxis}
\title{Function for computing basic matrix of rotation}
\usage{
Rx(theta)

Ry(theta)

Rz(theta)

Raxis(u, theta)
}
\arguments{
\item{theta}{numerical value in rad}

\item{u}{numeric 3 element vector describing axis of rotation}
}
\value{
3x3 matrix is returned.
}
\description{
Function for computing basic matrix of rotation
}
\details{
used in computing basic rotation matrices about x,y,z and u axis
}
\examples{
Rx(pi)
Ry(0.1)
Rz(-0.3)
Raxis(c(1,1,1), 0.1)

}
