% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shooterView3d.R
\name{shooterView3d}
\alias{shooterView3d}
\title{Function for visualizing positions of dronein 3D space}
\usage{
shooterView3d(
  pos = c(0, 0, 0),
  shooterPos = c(0, 0, 0),
  shooterA = diag(3),
  type = "points3D"
)
}
\arguments{
\item{pos}{3 element numeric vector describing position of drone "x","y","z"}

\item{shooterPos}{numeric 3 element vector describing position of shooter c(x,y,z)}

\item{shooterA}{matrix 3x3 describing orientation of shooter, default diag(3)}
}
\description{
Function for visualizing positions of dronein 3D space
}
\examples{
shooterView3d(pos = c(0,0,0), shooterPos = c(0,0,0), shooterA = diag(3))


}
