% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Function for computing skew symmetric matrix.}
\usage{
skew(vector)
}
\arguments{
\item{vector}{numerical vector c(a,b,c)}
}
\value{
3x3 matrix is returned.
}
\description{
Function for computing skew symmetric matrix.
}
\details{
Useful for computing cross product a x b = skew(a) \%*\% b
}
\examples{
skew (c(1,2,3))

}
