% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view3d.R
\name{view3d}
\alias{view3d}
\title{Function for visualizing positions of dronein 3D space}
\usage{
view3d(poses, shooterPos = c(0, 0, 0), shooterA = diag(3), type = "points3D")
}
\arguments{
\item{poses}{numeric array or dataframe with columns names "x","y","z"}

\item{shooterPos}{numeric 3 element vector c(x,y,z)}

\item{shooterA}{matrix 3x3 describing orientation of shooter, default diag(3)}

\item{type}{"points3D", "lines3D" default, "scatter3D"}
}
\value{
Numeric distance is returned.
}
\description{
Function for visualizing positions of dronein 3D space
}
\examples{
view3d(poses, type = "points3D")

}
